refactor dto
