+++ driver-service/src/main/java/io/voitovich/yura/driverservice/dto/mapper/DriverProfileMapper.java
@@ -1,7 +1,7 @@
 package io.voitovich.yura.driverservice.dto.mapper;
 
 
-import io.voitovich.yura.driverservice.dto.DriverProfileDto;
+import io.voitovich.yura.driverservice.dto.request.DriverProfileRequest;
 import io.voitovich.yura.driverservice.entity.DriverProfile;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
@@ -10,6 +10,6 @@ import org.mapstruct.factory.Mappers;
 public interface DriverProfileMapper {
 
     DriverProfileMapper INSTANCE = Mappers.getMapper(DriverProfileMapper.class);
-    DriverProfileDto toDto(DriverProfile profile);
-    DriverProfile toEntity(DriverProfileDto profileDto);
+    DriverProfileRequest toProfileRequest(DriverProfile profile);
+    DriverProfile toProfileEntity(DriverProfileRequest profileDto);
 }
